requirements.txt
docopt>=0.6.2
humanfriendly>=2.1
PyYAML>=3.12
